fixes errors

+++ install.sh
@@ -5,11 +5,12 @@
 ## linux安装
 #### tmux安装
 cd ~
+HOME=$(pwd)
 echo "now installing tmux"
 sudo apt install libevent ncurses
 sudo apt install tmux
 # 配置tmux支持快捷键与vim无缝切换
-if [ ! -e "~/.tmux.conf" ]; then
+if [ ! -e "$HOME/.tmux.conf" ]; then
     touch ~/.tmux.conf
 fi 
 cat>>~/.tmux.conf<<-"eof"
@@ -30,7 +31,8 @@ eof
 
 if [ ! -e "Python-3.8.8.tar.xz" ]; then
 	echo "Downloading Python Install Package."
-	sudo apt install -y zlib1g zlib1g-dev libffi-dev openssl libssl-dev
+»···sudo apt-get install build-essential
+	sudo apt install -y zlib1g zlib1g-dev libffi-dev openssl libssl-dev libbz2-dev liblzma-dev
 	wget https://www.python.org/ftp/python/3.8.8/Python-3.8.8.tar.xz
 fi
 if [ ! -e "/usr/local/python3.8/lib/python3.8/config-3.8-x86_64-linux-gnu" ]; then
@@ -57,50 +59,65 @@ fi
 cd ~
 #### 编译vim 支持Python和clipboard
 
-if [ ! -d "~/vim" ]; then
+if [ ! -d "$HOME/vim" ]; then
 	echo "Downloading Python Install Package."
 	cd ~
 	git clone https://github.com/vim/vim
 fi
-cd vim
-git pull && git fetch
 
-# 卸载旧版本vim
-sudo apt-get remove --purge vim-tiny vim vim-runtime vim-common vim-gui-common vim-nox
-# 安装依赖包
-sudo apt install ncurses-dev libgnome2-dev libgnomeui-dev libgtk2.0-dev libatk1.0-dev libbonoboui2-dev libcairo2-dev libx11-dev libxpm-dev
-libxt-dev lua5.1 liblua5.1-dev libperl-dev
-export LD_FLAGS="-rdynamic"
-#./configure --enable-multibyte --enable-python3interp=dynamic --with-python3-config-dir=/usr/local/python3.8/lib/python3.8/config-3.8-x86_64-linux-gnu --enable-cscopei --enable-gui=auto --with-features=huge --with-x --enable-fontset --enable-largefile --disable-netbeans --with-compiledby=SivanLaai --enable-fail-if-missing
-./configure --enable-multibyte --enable-python3interp=dynamic --with-python3-command=/usr/local/python3.8/bin/python3.8 \
---with-python3-config-dir=/usr/local/python3.8/lib/python3.8/config-3.8-x86_64-linux-gnu --enable-cscopei --enable-gui=auto --with-features=huge \
---enable-fontset --enable-largefile --disable-netbeans --with-compiledby=SivanLaai --enable-fail-if-missing
-make
-sudo make install
+
 cd ~
+output=$(vim --version | grep "SivanLaai")
+if [ ! -n "$output" ]
+then
+	cd ~/vim
+	git pull && git fetch
+	
+	# 卸载旧版本vim
+	sudo apt-get remove --purge vim-tiny vim vim-runtime vim-common vim-gui-common vim-nox
+	# 安装依赖包
+	sudo apt install ncurses-dev libgnome2-dev libgnomeui-dev libgtk2.0-dev libatk1.0-dev libbonoboui2-dev libcairo2-dev libx11-dev libxpm-dev
+	libxt-dev lua5.1 liblua5.1-dev libperl-dev
+	export LD_FLAGS="-rdynamic"
+	#./configure --enable-multibyte --enable-python3interp=dynamic --with-python3-config-dir=/usr/local/python3.8/lib/python3.8/config-3.8-x86_64-linux-gnu --enable-cscopei --enable-gui=auto --with-features=huge --with-x --enable-fontset --enable-largefile --disable-netbeans --with-compiledby=SivanLaai --enable-fail-if-missing
+	./configure --enable-multibyte --enable-python3interp=dynamic --with-python3-command=/usr/local/python3.8/bin/python3.8 \
+	--with-python3-config-dir=/usr/local/python3.8/lib/python3.8/config-3.8-x86_64-linux-gnu --enable-cscopei --enable-gui=auto --with-features=huge \
+	--enable-fontset --enable-largefile --disable-netbeans --with-compiledby=SivanLaai --enable-fail-if-missing
+	make
+	sudo make install
+	cd ~
+else
+	echo "vim 8.2 has been installed successfully."
+fi
 
 # 编译nodejs,插件coc需要nodejs的功能
 # 安装nodejs, 插件coc.vim会用到这个软件
+cd ~
 if [ ! -e "node-v14.17.1-linux-x64.tar.xz" ]; then
 	echo "Downloading node-v14.17.1 Install Package."
 	wget https://nodejs.org/dist/v14.17.1/node-v14.17.1-linux-x64.tar.xz
 fi
-tar -xf node-v14.17.1-linux-x64.tar.xz
-echo "installing nodejs"
-sudo mv node-v14.17.1-linux-x64 /usr/local/nodejs 
+if [ ! -e "/usr/local/nodejs/bin/node" ]; then
+	tar -xf node-v14.17.1-linux-x64.tar.xz
+	echo "installing nodejs"
+	sudo mv node-v14.17.1-linux-x64 /usr/local/nodejs 
+fi
 
 #添加nodejs环境变量
 bash_file="~/.bash_profile"
-echo ${bash_file}
-if [ ! -e "~/.bash_profile" ]; then
-    touch ~/.bash_profile
+if [ ! -e "$HOME/.bash_profile" ]; then
+    touch $HOME/.bash_profile
 fi 
-cat>>~/.bash_profile<<-"eof"
-#config nodejs env path
-VERSION=v14.17.1
-DISTRO=linux-x64
-export PATH=/usr/local/nodejs/bin:$PATH
-eof
+
+output=$(cat ~/.bash_profile | grep "config nodejs env path")
+if [ ! -n "$output" ]; then
+	cat>>$HOME/.bash_profile<<-"eof"
+	#config nodejs env path
+	VERSION=v14.17.1
+	DISTRO=linux-x64
+	export PATH=/usr/local/nodejs/bin:$PATH
+	eof
+fi
 
 #安装ctags，主要用来索引
 #安装依赖
@@ -112,12 +129,14 @@ sudo apt install \
     libjansson-dev \
     libyaml-dev \
 	libxml2-dev
-if [ ! -d "~/universal-ctags" ]; then
+if [ ! -d "$HOME/ctags" ]; then
 	echo "Downloading universal-ctags Install Package."
-	git clone https://github.com/universal-ctags/ctags.git
+	output=$(git clone https://github.com/universal-ctags/ctags.git)
+	echo $output
 fi
-if [ ! -d "/usr/local/ctags" ]; then
-	echo "Downloading universal-ctags Install Package."
+
+if [ ! -e "/usr/local/bin/ctags" ]; then
+	echo "compiling universal-ctags."
 	cd ctags
 	./autogen.sh
 	./configure --prefix=/usr/local # defaults to /usr/local
@@ -130,20 +149,17 @@ cd ~
 # 安装vim-plug
 #curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
 #    https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-if [ ! -d "~/vim-plug" ]; then
+if [ ! -e "$HOME/.vim/autoload/plug.vim" ]; then
 	echo "Downloading vim-plug Install Package."
 	git clone https://github.com/junegunn/vim-plug.git
+	# 如果安装失败的话，可能就需要修改plug.vim
+	# find ~ -name 'plug.vim' | xargs perl -pi -e 's|https://git::@github.com/%s.git|git@github.com:%s.git|g'
+	mkdir -p $HOME/.vim/autoload
 fi
-# 如果安装失败的话，可能就需要修改plug.vim
-# find ~ -name 'plug.vim' | xargs perl -pi -e 's|https://git::@github.com/%s.git|git@github.com:%s.git|g'
-if [ ! -d "~/.vim/autoload" ]; then
-	echo "~/.vim/autoload not exsits, now create"
-	mkdir -p ~/.vim/autoload
-fi
-cd ~/vim-light-workshop
-cp -rf vimrc ~/.vimrc
-cp -rf gvim/colors ~/.vim/colors
-cp -rf ~/vim-plug/plug.vim ~/.vim/autoload/plug.vim
+cd $HOME/vim-light-workshop
+cp -rf vimrc $HOME/.vimrc
+cp -rf gvim/colors $HOME/.vim/colors
+cp -rf $HOME/vim-plug/plug.vim $HOME/.vim/autoload/plug.vim
 
 #### zsh安装
 if ! [ -x "$(command -v zsh)" ]; then
@@ -154,40 +170,29 @@ else
 	echo 'zsh is installed'
 fi
 #安装oh-my-zsh
-if ! [ ! -d "~/.oh-my-zsh" ]; then
+if [ ! -d "$HOME/ohmyzsh" ]; then
 	echo 'installing oh-my-zsh now.' >&2
 	cd ~
-	git clone https://github.com/ohmyzsh/ohmyzsh.git
+	output=$(git clone https://github.com/ohmyzsh/ohmyzsh.git)
+	echo $output
 else
 	echo 'oh-my-zsh is installed, now reinstall oh-my-zsh'
-	rm -rf ~/.oh-my-zsh
-	rm -rf ~/.zshrc
+	rm -rf $HOME/.oh-my-zsh
+	rm -rf $HOME/.zshrc
 fi
 
-cd ~/ohmyzsh/tools
-./install.sh
+cd ~
+cd $HOME/ohmyzsh/tools
 #install theme
+./install.sh
 git clone --depth=1 https://gitee.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
 sed -i 's#"robbyrussell"#"powerlevel10k/powerlevel10k"#g' ~/.zshrc
 #激活环境bash变量
-echo 'source ~/.bash_profile'>>~/.zshrc
-cd ~/vim-light-workshop
-
-#### 安装autojump
-if ! [ ! -d "~/.autojump" ]; then
-	echo 'installing autojump now.' >&2
-	git clone https://github.com/joelthelion/autojump.git
-	cd autojump
-	./install.py
-	cat>>~/.zshrc<<-"eof"
-	[[ -s /home/laixinhua/.autojump/etc/profile.d/autojump.sh ]] && source /home/laixinhua/.autojump/etc/profile.d/autojump.sh
-	autoload -U compinit && compinit -u
-	eof
-	cd ~
-	sudo rm -rf autojump
-else
-	echo 'autojump is installed'
+output=$(cat ~/.zshrc | grep "bash_profile")
+if [ ! -n "$output" ]; then
+	echo 'source ~/.bash_profile'>>~/.zshrc
 fi
+
 cd ~
 
 cat<<"eof"
